requirements.txt
streamlit==1.35.0
langchain==0.2.6
langchain-community==0.2.6
langchain-huggingface==0.0.3
pymupdf==1.24.5
sentence-transformers==3.0.1
faiss-cpu==1.8.0
groq==0.9.0
python-dotenv==1.0.1
torch==2.3.1
pytesseract==0.3.10
pdf2image==1.17.0